WIP - remove redundant pd proxy mem obj. Consolidate transfer ctx
